Put gapped series logic in a shared hook
